requirements.txt
math
numpy
pandas
openpyxl
matplotlib
seaborn
pillow
scikit-learn
tensorflow